[FIX] Шульпин Илья. Вариант 23. Технология TBB. Построение выпуклой оболочки – проход Джарвиса. (#592)
**Описание ошибки**
Ошибка была вызвана из-за того, что при проверки поворта программа могла
посчитать точки коллиниарными в тех случаях, когда поворот действительно
был

**Пример ошибки**
val=(q.y−p.y)×(r.x−q.x)−(q.x−p.x)×(r.y−q.y)

A -- (0, 0)
B -- (1, 1.0000000001)
C -- (2, 2)

val=(1.0000000001−0)×(2−1)−(1−0)×(2−1.0000000001)
=(1.0000000001)×(1)−(1)×(0.9999999999) = =1.0000000001−0.9999999999 =
2×10^(-10)

Это ниже, чем пороговое значение, при котором точки не считаются
коллиниарными. Из-за этого точка С рассматривается программой как
колиниарная и не включается в оболочку, хотя должна

**Что исправлено**
В целях безопасности работы алгоритма все координаты точек для тестов на
случайных данных сделаны целочисленными

+++ tasks/tbb/shulpin_i_jarvis_passage/src/test_modules.cpp
@@ -222,9 +222,9 @@ std::vector<shulpin_i_jarvis_tbb::Point> shulpin_tbb_test_module::GenerateRandom
   std::uniform_int_distribution<int> dist(-10000, 10000);
 
   for (size_t i = 0; i < num_points; ++i) {
-    double x = dist(gen);
-    double y = dist(gen);
-    points.emplace_back(x, y);
+    int x = dist(gen);
+    int y = dist(gen);
+    points.emplace_back(static_cast<double>(x), static_cast<double>(y));
   }
 
   return points;
